requirements.txt
dvc==3.48.4
mlflow==2.12.1
scikit-learn==1.4.2
pandas==1.5.3
numpy==1.24.3
feast==0.36.0
psycopg2-binary==2.9.9
sqlalchemy==1.4.52
joblib==1.4.0
pyyaml==6.0.1
pyarrow==15.0.2
pathspec==0.11.2